che hochy to i vorochy. moya vetka, ponyal?

+++ src/main/java/com/peter/kist/repository/MarkRepository.java
@@ -3,5 +3,5 @@ package com.peter.kist.repository;
 import com.peter.kist.model.Mark;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarkRepository extends JpaRepository<Mark, Long> {
+public interface MarkRepository extends JpaRepository<Mark, Integer> {
 }
